+++ lifelines/utils/__init__.py
@@ -10,6 +10,7 @@ from textwrap import dedent
 import numpy as np
 from scipy.linalg import solve
 from scipy import stats
+from scipy.integrate import quad, trapz
 import pandas as pd
 
 from lifelines.utils.concordance import concordance_index
@@ -19,6 +20,7 @@ __all__ = [
     "qth_survival_times",
     "qth_survival_time",
     "median_survival_times",
+    "restricted_mean_survival_time",
     "survival_table_from_events",
     "group_survival_table_from_events",
     "survival_events_from_table",
@@ -169,7 +171,7 @@ def qth_survival_time(q, survival_function, cdf=False):
     --------
     qth_survival_times, median_survival_times
     """
-    if type(survival_function) is pd.DataFrame:  # pylint: disable=unidiomatic-typecheck
+    if isinstance(survival_function, pd.DataFrame):
         if survival_function.shape[1] > 1:
             raise ValueError(
                 "Expecting a dataframe (or series) with a single column. Provide that or use utils.qth_survival_times."
@@ -187,8 +189,86 @@ def qth_survival_time(q, survival_function, cdf=False):
     return v
 
 
-def median_survival_times(density_or_survival_function, left_censorship=False):
-    return qth_survival_times(0.5, density_or_survival_function, cdf=left_censorship)
+def median_survival_times(model_or_density_or_survival_function, left_censorship=False):
+    """
+    Compute the median survival time of survival function(s).
+
+    Parameters
+    -----------
+
+    model_or_density_or_survival_function: lifelines model or DataFrame
+        This can be a univariate lifelines model, or a DataFrame of one or more survival functions.
+
+    """
+    import lifelines
+
+    if isinstance(model_or_density_or_survival_function, pd.DataFrame):
+        return qth_survival_times(0.5, model_or_density_or_survival_function, cdf=left_censorship)
+    elif isinstance(model_or_survival_function, lifelines.fitters.UnivariateFitter):
+        return model.median_
+    else:
+        raise ValueError("Can't compute median surviva time of object %s" % model_or_survival_function)
+
+
+def restricted_mean_survival_time(model_or_survival_function, t=np.inf):
+    r"""
+    Compute the restricted mean survival time, RMST, of a survival function. This is defined as
+
+    .. math::  \text{RMST}(t) = \int_0^t S(\tau) d\tau
+
+    For reason why we use an upper bound and not always :math:`\infty` is because the tail of a survival
+    function has high variance and strongly effects the RMST.
+
+    Parameters
+    -----------
+
+    model_or_survival_function: lifelines model or DataFrame
+        This can be a univariate model, or a pandas DataFrame. The former will provide a more accurate estimate however.
+    t: float
+        The upper limit of the integration in the RMST.
+
+    Example
+    --------
+
+    >>> from lifelines import KaplanMeierFitter, WeibullFitter
+    >>> from lifelines.utils import restricted_mean_survival_time
+    >>>
+    >>> kmf = KaplanMeierFitter().fit(T, E)
+    >>> restricted_mean_survival_time(kmf, t=3.5)
+    >>> restricted_mean_survival_time(kmf.survival_function_, t=3.5)
+    >>>
+    >>> wf = WeibullFitter().fit(T, E)
+    >>> restricted_mean_survival_time(wf)
+    >>> restricted_mean_survival_time(wf.survival_function_)
+
+    References
+    -------
+    https://bmcmedresmethodol.biomedcentral.com/articles/10.1186/1471-2288-13-152#Sec27
+
+    """
+    import lifelines
+
+    if isinstance(model_or_survival_function, pd.DataFrame):
+        warnings.warn(
+            "Approximating RMST using the precomputed survival function. You likely will get a more accurate estimate if you provide the fitted Model instead of the survival function.",
+            UserWarning,
+        )
+        sf = model_or_survival_function.loc[:t]
+        sf = sf.append(pd.DataFrame([1], index=[0], columns=sf.columns)).sort_index()
+        return trapz(y=sf.values[:, 0], x=sf.index)
+    elif isinstance(model_or_survival_function, lifelines.fitters.UnivariateFitter):
+        # lifelines model
+        model = model_or_survival_function
+        # if KM, we can compute exactly
+        if isinstance(model, lifelines.KaplanMeierFitter):
+            sf = model.survival_function_.loc[:t]
+            sf = sf.append(pd.DataFrame([model.predict(t)], index=[t], columns=sf.columns)).sort_index()
+            sf = sf.reset_index()
+            return (sf["index"].diff().shift(-1) * sf[model._label]).sum()
+        else:
+            return quad(model.survival_function_at_times, 0, t)[0]
+    else:
+        raise ValueError("Can't compute RMST of object %s" % model_or_survival_function)
 
 
 def group_survival_table_from_events(
@@ -311,6 +391,8 @@ def survival_table_from_events(
     intervals=None,
 ):  # pylint: disable=dangerous-default-value,too-many-locals
     """
+    Create a survival table from right-censored dataset.
+
     Parameters
     ----------
     death_times: (n,) array
@@ -466,6 +548,10 @@ def survival_events_from_table(survival_table, observed_deaths_col="observed", c
     >>> E = np.array([ 1.,  0.,  1.,  1.,  0.,  0.])
     >>> W = np.array([ 1,  1,  1,  1,  1,  1])
 
+    See Also
+    --------
+    ``survival_table_from_events``
+
     """
     T_ = []
     E_ = []
@@ -541,7 +627,7 @@ def datetimes_to_durations(
 
     T = (end_times_ - start_times_).values.astype(freq_string).astype(float)
     if (T < 0).sum():
-        warnings.warn("Warning: some values of start_times are after end_times")
+        warnings.warn("Warning: some values of start_times are after end_times", UserWarning)
     return T, C.values
 
 
@@ -1006,7 +1092,7 @@ def check_nans_or_infs(df_or_array):
             """Attempting to convert an unexpected datatype '%s' to float. Suggestion: 1) use `lifelines.utils.datetimes_to_durations` to do conversions or 2) manually convert to floats/booleans."""
             % df_or_array.dtype
         )
-        warnings.warn(warning_text)
+        warnings.warn(warning_text, UserWarning)
         try:
             infs = np.isinf(df_or_array.astype(float))
         except: